+++ doc/CHANGELOG.md
@@ -1,5 +1,11 @@
 # yetibot.core change log
 
+## 0.4.44 - Unreleased
+
+- Use ping/pong events in Slack to monitor the connection
+- Expose connection latency metrics on GraphQL `:adapters` resolver
+- Upgrade to `yetibot-dashboard 0.6.6`
+
 ## 0.4.43 - 10/15/2018
 
 - Improve logging in interpreter